requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.6,<3.0
python-dotenv>=1.0,<2.0
openai>=1.30.0,<2.0
httpx>=0.27,<1.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
pypdf>=4.2,<6.0
python-docx>=1.1,<2.0
